v.1.0.1
1. Newly added pages, elements and tags were being lost if cropping
done prior to saving item. Updated to include elements being newly
added in lists returned for pages, elements and tags.

2. Removal of tag not working when attempting to remove last tag.
Updated to use input and not select in html form for input of tags, so
that when zero items a null value is returned.

3. Publishing items with multiple images was slow because all images
for the item required for web site in all required formats were
recreated. Updated to create web images only when an image is newly
added or cropping has changed, or doing an image recreate.

4. Sequencing of images tidied up.

5. Updated to create thumbnail images correctly when not set to 1:1.